Add support for nested calls
The PCM supports calling functions from other functions (up to two
levels deep), so this emulator should support that, too.

This code is kind of a mess. Rather than setting r11 independently and
passing the delayed PC and call state as tuple values to
ExecState::DelaySlot(), it may be better to bundle those values into a
"next state" struct, which would include r11 if it needs to be updated,
as well as the delayed PC and call state.