Use a trait to implement register and memory hooks
Using a trait instead of passing functions as parameters enables the
compiler to optimize the emulator core at compile time (e.g., by
inlining the functions or eliminating branches if a hook isn't used).
Since these functions are executed in very hot paths in the emulator
core (register and memory access), these optimizations can result in
significant speedups. Testing with a loop program showed a 10% to 25%
increase in performance.